API review update - FolderNames
