+++ src/Microsoft.DotNet.Wpf/src/PresentationFramework/ref/PresentationFramework.cs
@@ -17,11 +17,7 @@ namespace Microsoft.Win32
         public System.Collections.Generic.IList<Microsoft.Win32.FileDialogCustomPlace> CustomPlaces { get { throw null; } set { } }
         public bool DereferenceLinks { get { throw null; } set { } }
         public string InitialDirectory { get { throw null; } set { } }
-        public string FileName { get { throw null; } set { } }
-        public string[] FileNames { get { throw null; } }
         public bool RestoreDirectory { get { throw null; } set { } }
-        public string SafeFileName { get { throw null; } }
-        public string[] SafeFileNames { get { throw null; } }
         public string Title { get { throw null; } set { } }
         public bool ValidateNames { get { throw null; } set { } }
         public event System.ComponentModel.CancelEventHandler FileOk { add { } remove { } }
@@ -38,9 +34,13 @@ namespace Microsoft.Win32
         public bool CheckFileExists { get { throw null; } set { } }
         public bool CheckPathExists { get { throw null; } set { } }
         public string DefaultExt { get { throw null; } set { } }
+        public string FileName { get { throw null; } set { } }
+        public string[] FileNames { get { throw null; } }
         public string Filter { get { throw null; } set { } }
         public int FilterIndex { get { throw null; } set { } }
         public override void Reset() { }
+        public string SafeFileName { get { throw null; } }
+        public string[] SafeFileNames { get { throw null; } }
         public override string ToString() { throw null; }
     }
     public sealed partial class FileDialogCustomPlace
@@ -83,8 +83,12 @@ namespace Microsoft.Win32
     public sealed partial class OpenFolderDialog : Microsoft.Win32.CommonItemDialog
     {
         public OpenFolderDialog() { }
+        public string FolderName { get { throw null; } set { } }
+        public string[] FolderNames { get { throw null; } }
         public bool Multiselect { get { throw null; } set { } }
         public override void Reset() { }
+        public string SafeFolderName { get { throw null; } }
+        public string[] SafeFolderNames { get { throw null; } }
     }
     public sealed partial class SaveFileDialog : Microsoft.Win32.FileDialog
     {
